style: add utils
