+++ src/main/java/com/example/carrentalproject/CarRentalApplication.java
@@ -1,17 +1,9 @@
 package com.example.carrentalproject;
 
 
-import com.example.carrentalproject.domain.Role;
-import com.example.carrentalproject.dto.UserDto;
-import com.example.carrentalproject.repo.RoleRepository;
-import com.example.carrentalproject.service.UserService;
-import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-import java.util.ArrayList;
 
 
 @SpringBootApplication
